[임선혁] 인기도 기능 80%

+++ ConnectionPool/src/main/java/com/kh/cntp/battle/model/vo/ResultHistory.java
@@ -10,5 +10,5 @@ public class ResultHistory {
 	private int victory;
 	private int defeat;
 	private int record;
-	private float winRate;
+	private double winRate;
 }
